requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.1
matplotlib==3.8.3
seaborn==0.13.2
Pillow==10.2.0